requirement.txt
pymupdf
streamlit
pandas
pafy
plotly
pymysql
streamlit-tags
Pillow
